requirements.txt
Django>=1.8,<2.0
PyMySQL
djangorestframework
django-filter